# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Clinic booking API (learning)
  description: |
    Hand-maintained OpenAPI spec. Update this file when routes change; UI at `/api/docs`.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local (default PORT from `.env`)
...
paths:
...
  /api/v1/doctors/me/slots:
    get:
      tags: [Doctors]
      summary: List all slots for the authenticated doctor’s linked record
      operationId: listMyDoctorSlots
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All slots for `doctorRecordId` (available and booked)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Slot"
        "401":
          description: Missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "403":
          description: Not a doctor or profile not linked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
    post:
      tags: [Doctors]
      summary: Create a slot for the authenticated doctor’s linked directory record
      operationId: createMyDoctorSlot
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSlotRequest"
      responses:
        "201":
          description: Slot created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Slot"
        "400":
          description: Validation error (missing times, invalid datetimes, or end not after start)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "401":
          description: Missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "403":
          description: Not a doctor or profile not linked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "404":
          description: Doctor record not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "409":
          description: New slot overlaps another opening for the same doctor (`SLOT_OVERLAP`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/v1/doctors/me/slots/{slotId}:
    delete:
      tags: [Doctors]
      summary: Delete a diary opening when it has no active visits (authenticated doctor)
      operationId: deleteMyDoctorSlot
      security:
        - bearerAuth: []
      parameters:
        - name: slotId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "204":
          description: Slot removed
        "400":
          description: Invalid slot id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "401":
          description: Missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "403":
          description: Not owner doctor or profile not linked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "404":
          description: Slot not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "409":
          description: Slot still has a pending or confirmed appointment (`SLOT_IN_USE`); only cancelled/rejected history is removed with the slot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/v1/doctors/{id}:
    get:
      tags: [Doctors]
      summary: Get doctor by id
      operationId: getDoctorById
      parameters:
        - $ref: "#/components/parameters/DoctorId"
      responses:
        "200":
          description: Doctor row
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"
        "404":
          description: Doctor not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/v1/doctors/{id}/slots:
    get:
      tags: [Doctors]
      summary: List available slots for a doctor
      operationId: listDoctorSlots
      parameters:
        - $ref: "#/components/parameters/DoctorId"
      responses:
        "200":
          description: Available slots only
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Slot"
        "400":
          description: Invalid doctor id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "404":
          description: Doctor not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
    post:
      tags: [Doctors]
      summary: Create a slot (authenticated doctor for own doctor id)
      operationId: createDoctorSlot
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/DoctorId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSlotRequest"
      responses:
        "201":
          description: Slot created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Slot"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "403":
          description: Not owner doctor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "404":
          description: Doctor not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "409":
          description: New slot overlaps another opening for the same doctor (`SLOT_OVERLAP`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
...
  /api/v1/appointments/doctor:
    get:
      tags: [Appointments]
      summary: List appointments on slots owned by the authenticated doctor
      operationId: listDoctorAppointments
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Appointments for doctor slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "403":
          description: Not a doctor or doctor not linked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Use Authorization header — `Bearer <token>` from `/api/v1/auth/login`
  parameters:
    DoctorId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Doctor primary key (numeric string in path)
  schemas:
...
    Doctor:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        specialty:
          type: string
    Slot:
      type: object
      properties:
        id:
          type: integer
        doctorId:
          type: integer
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        isAvailable:
          type: integer
          description: SQLite stores 0/1
    Appointment:
      type: object
      properties:
        id:
          type: integer
        slotId:
          type: integer
        patientId:
          type: integer
        status:
          type: string
          example: pending
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          nullable: true
...
    CreateSlotRequest:
      type: object
      required: [startTime, endTime]
      properties:
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        isAvailable:
          type: boolean
          default: true
    ErrorBody:
      type: object
      properties:
        errorCode:
          type: string
        message:
          type: string
        requestId:
          type: string
          format: uuid